fix: docstring
